feat: lv3 비밀번호 추가 완료

+++ src/main/java/com/example/developedscheduler/service/UserService.java
@@ -19,7 +19,7 @@ public class UserService {
     private final UserRepository userRepository;
     private final ScheduleRepository scheduleRepository;
 
-    public UserResponseDto signUpUser(String name, String address) {
+    public UserResponseDto signUpUser(String name, String email, String password) {
 
         //사용자가 작성한 이름이 db에 있는지 확인 boolean 반환
         boolean existsUserByName = userRepository.existsUserByName(name);
@@ -30,12 +30,12 @@ public class UserService {
         }
 
         //없다면 새로운 유저 생성
-        User user = new User(name, address);
+        User user = new User(name, email, password);
 
         //유저 저장
         User savedUser = userRepository.save(user);
 
-        return new UserResponseDto(savedUser.getId(),savedUser.getName(),savedUser.getAddress());
+        return new UserResponseDto(savedUser.getId(),savedUser.getName(),savedUser.getEmail());
     }
 
 
@@ -53,11 +53,11 @@ public class UserService {
         //특정 아이디로 유저 조회
         User foundUser = userRepository.findByIdOrElseThrow(id);
 
-        return new UserResponseDto(foundUser.getId(), foundUser.getName(), foundUser.getAddress());
+        return new UserResponseDto(foundUser.getId(), foundUser.getName(), foundUser.getEmail());
     }
 
     @Transactional
-    public UserResponseDto updateUser(Long id, String name, String address) {
+    public UserResponseDto updateUser(Long id, String name, String email, String oldPassword, String newPassword) {
 
         User foundUser = userRepository.findByIdOrElseThrow(id);
 
@@ -66,22 +66,41 @@ public class UserService {
             name = foundUser.getName();
         }
 
-        if(address == null) {
-            address = foundUser.getAddress();
+        if(email == null) {
+            email = foundUser.getEmail();
         }
 
-        //수정완료
-        foundUser.updateUser(name,address);
+        if(oldPassword == null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,"비밀번호를 입력해주세요");
+        }
+
+        if(foundUser.getPassword().equals(oldPassword)) {
+            if(newPassword == null) {
+                foundUser.updateUser(name,email,oldPassword);
+            } else {
+                foundUser.updateUser(name,email,newPassword);
+            }
+        } else {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,"비밀번호가 틀렸습니다.");
+        }
 
         return UserResponseDto.toDto(foundUser);
     }
 
 
     @Transactional
-    public void deleteUser(Long id) {
+    public void deleteUser(Long id,String password) {
 
         //유저 있는지 확인
         User foundUser = userRepository.findByIdOrElseThrow(id);
+
+        if(password == null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,"비밀번호를 입력해주세요");
+        }
+
+        if(!password.equals(foundUser.getPassword())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,"비밀번호가 틀렸습니다. 재입력해주세요.");
+        }
 
         //유저가 있다면 그 유저의 아이디로 작성된 모든 일정 삭제
         scheduleRepository.deleteSchedulesByUser_Id(id);
